Aid IDE code linking in V3AstNodes.h
