+++ src/V3AstNodes.h
@@ -19,6 +19,7 @@
 
 #ifndef VERILATOR_V3AST_H_
 #error "Use V3Ast.h as the include"
+#include "V3Ast.h"  // This helps code analysis tools pick up symbols in V3Ast.h
 #endif
 
 //######################################################################
@@ -9153,5 +9154,7 @@ public:
 };
 
 //######################################################################
+
+#include "V3AstInlines.h"
 
 #endif  // Guard
